pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unreal-engine-agentic-control"
version = "0.1.0"
description = "Natural language control of the Unreal Engine Editor via AI agents"
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.25.0",
    "fastmcp==2.14.5",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.setuptools.packages.find]
include = ["agents*", "mcp_server*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
